Add new directory, with file optional functions Fixed of methods _append of influx_of_water.py
