+++ explorer/__init__.py
@@ -1,2 +1,2 @@
-# pytank/graphs/__init__.py
+# pytank/explorer/__init__.py
 from .explore_data_analysis import *
